docker/purchase_service/requirements.txt
flask
psycopg2-binary
flask_jwt_extended
flask_sqlalchemy
requests
flask_cors
flasgger
waitress